fix irseeker sensor support

+++ lib/irseeker/ev3wrapirseeker.hpp
@@ -7,8 +7,8 @@ namespace Ev3Wrap {
 class IrSeeker : private ev3dev::irseeker_sensor {
     public:
         static IrSeeker bind(ev3dev::address_type addr = ev3dev::INPUT_AUTO);
-        float getDirectionAC() {return this->ac();}
-        float getDirectionDC() {return this->dc();}
+        int getDirectionAC() {return this->ac();}
+        int getDirectionDC() {return this->dc();}
         std::tuple<int, int, int, int, int, int> getAllAC(){return this->ac_all();}
         std::tuple<int, int, int, int, int, int, int> getAllDC(){return this->dc_all();}
         // future update
